Update Medium.java
Given an array of integers, print a sum triangle from it such that the first level(the bottom level in triangular fashion) has all array elements. From then, at each level, the number of elements is one less than the previous level and elements at the level is the sum of consecutive two elements in the previous level.